Implemenation of Authorization and relationships

+++ Controllers/expenseController.js
@@ -1,10 +1,16 @@
 const path = require('path');
 const ExpenseData = require('../Models/ExpenseModel');
+
+exports.getExpenseMainHomePage = (req, res) => {
+    res.sendFile(path.join(__dirname, "..", "Views", "mainHome.html"));
+};
+
 exports.getExpenseMainPage = (req, res) => {
     res.sendFile(path.join(__dirname, "..", "Views", "expenseMain.html"));
 };
 exports.addExpense = async (req, res) => {
     const body = req.body;
+    const id = req.user.id;
     const expenseAmount = body.ExpenseAmount;
     const description = body.ExpenseDesc;
     const expenseType = body.ExpenseType;
@@ -12,7 +18,8 @@ exports.addExpense = async (req, res) => {
         await ExpenseData.create({
             expenseAmount: expenseAmount,
             description: description,
-            expenseType: expenseType
+            expenseType: expenseType,
+            userDatumId: id
         });
         res.status(201).json({ data: 'success' })
     } catch (err) {
@@ -26,7 +33,8 @@ exports.getExpensesViewPage = (req, res) => {
 
 exports.getExpensesData = async (req, res) => {
     try {
-        const result = await ExpenseData.findAll();
+        const id = req.user.id;
+        const result = await ExpenseData.findAll({ where: { userDatumId: id } });
         res.json(result);
     } catch (err) {
         res.status(500).json({ data: 'error' });
@@ -37,7 +45,9 @@ exports.getExpensesData = async (req, res) => {
 exports.deleteExpenseData = async (req, res) => {
     try {
         const id = req.body.id
-        await ExpenseData.destroy({ where: { id: id } });
+        const userid = req.user.id;
+        console.log(userid)
+        await ExpenseData.destroy({ where: { id: id, userDatumId: userid } });
         res.redirect('/expense/viewExpenses');
     }
     catch (err) {
@@ -49,11 +59,12 @@ exports.deleteExpenseData = async (req, res) => {
 exports.updateExpense = async (req, res) => {
     const body = req.body;
     const id = body.id;
+    const userid = req.user.id;
     const expenseAmount = body.data.ExpenseAmount;
     const description = body.data.ExpenseDesc;
     const expenseType = body.data.ExpenseType;
     try {
-        const expenseData = await ExpenseData.findByPk(id);
+        const expenseData = await ExpenseData.findOne({ where: { id: id, userDatumId: userid } });
         expenseData.expenseAmount = expenseAmount;
         expenseData.description = description;
         expenseData.expenseType = expenseType;
